requirements.txt
streamlit==1.31.0
PyPDF2==3.0.1
sentence-transformers==2.7.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-huggingface==0.1.0
faiss-cpu==1.8.0
groq==0.9.0
python-dotenv==1.0.1
httpx==0.27.2